requirements.txt
click>=8.0.0
openai>=1.0.0
anthropic>=0.20.0
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.28.0
aiohttp>=3.8.0
asyncio>=3.4.3
rich>=13.0.0
pillow>=10.0.0
pyperclip>=1.8.0